+++ CHANGELOG.md
@@ -18,11 +18,16 @@
 ### :rocket: New Feature
 - `preview`
   - [#1356](https://github.com/wix-incubator/rich-content/pull/1356) support video in gallery
+  - [#1369](https://github.com/wix-incubator/rich-content/pull/1369) preview content examples in storybook
+- `storybook`
+  - [#1369](https://github.com/wix-incubator/rich-content/pull/1369) preview content examples
 - `giphy`
   - [#1358](https://github.com/wix-incubator/rich-content/pull/1358) gi preview modal empty state
 - `ricos-viewer`
   - [#1249](https://github.com/wix-incubator/rich-content/pull/1249) preview strategy
 ### :bug: Bug Fix
+- `gallery`
+  - [#1362](https://github.com/wix-incubator/rich-content/pull/1362) fix rtl
 - `preview`
   - [#1361](https://github.com/wix-incubator/rich-content/pull/1361) readMore - removed automatic scroll to top page
   - [#1366](https://github.com/wix-incubator/rich-content/pull/1366) support multiple vids, gifs & wix media